--- .github/workflows/main.yml
@@ -1,18 +1,0 @@
-name: eslint action
-on: 
-  push: 
-    branches: 
-      - development
-jobs:
-    eslint: 
-      runs-on: ubuntu-latest
-      steps: 
-        - name:  checkout action
-          uses : actions/checkout@v3
-        - name: install dependecies
-          run: npm install 
-        - name: init eslint
-          id: eslint
-          run: npm run lint
-        - name: mergear branch development com produção
-          run: git status
