+++ FrontEnd/src/App.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import NavBar from "./Components/navBar";
+
+function App() {
+    return (
+        <main>
+            <NavBar/>
+        </main>
+    );
+}
+
+export default App;
