--- src/react-app-env.d.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/triple-slash-reference
-/// <reference types="react-scripts" />
